+++ src/test/java/im/toduck/domain/routine/domain/service/RoutineServiceTest.java
@@ -1,6 +1,6 @@
 package im.toduck.domain.routine.domain.service;
 
-import static im.toduck.fixtures.UserFixtures.*;
+import static im.toduck.fixtures.user.UserFixtures.*;
 import static org.assertj.core.api.SoftAssertions.*;
 
 import java.time.DayOfWeek;
