+++ src/test/java/im/toduck/domain/routine/persistence/repository/RoutineRecordRepositoryTest.java
@@ -2,7 +2,7 @@ package im.toduck.domain.routine.persistence.repository;
 
 import static im.toduck.fixtures.RoutineFixtures.*;
 import static im.toduck.fixtures.RoutineRecordFixtures.*;
-import static im.toduck.fixtures.UserFixtures.*;
+import static im.toduck.fixtures.user.UserFixtures.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.SoftAssertions.*;
